Refactor code for future usage szenarios
- Encapsulate faults into a fault module with single files for every fault type
- Remove first parts of fault knowledge from cpu struct
